requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.21.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==3.0.1
mysqlclient==1.4.2.post1
passlib==1.7.1
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.6
webargs==5.4.0
Werkzeug==0.15.5
